Add <MentionsInput /> component
